% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ciccaba_App.R
\name{Ciccaba_App}
\alias{Ciccaba_App}
\title{Deploy a Shiny app to input daily monitoring recordings.}
\usage{
Ciccaba_App()
}
\value{
Launches the Shiny app
}
\description{
Deploy a Shiny app to input daily monitoring recordings.
}
\examples{

Ciccaba_App()
}
